Specification Revision & Myers program updates
The specification document was substantial revised to make the
structure of the document clearer.  Also the number of primary types
was reduced to just 5 with a number of refinement/corrections to the
alignment and link data types.

Top level Makefile was also revised to make everything in Myers and
Durbin subdirectories and place them in a common /bin so users can see
what tools are currently available.

Minor fixes to some of Myers tools.  VGP cloud is now functional.